refactor(api): move find paginated filtered users usecase
in src/identity-access-management
